Make the ROM's own main the port's boot spine by default
Lane BOOTR1, run link100. Lane BOOTSCOUT built this behind SM64DS_ROM_MAIN and
left the gate off because a measurement should not change what runs. The
measurement came back clean, so the default moves: unset, or any value other
than "0", now runs src/main.c; SM64DS_ROM_MAIN=0 keeps the transcribed head in
hal/boot_os.cpp.

WHY THE ROM'S MAIN IS THE FAITHFUL BOOT, re-derived rather than taken on trust.
main at arm9 0x02007000 was read out of extracted/arm9_dec.bin and matches
src/main.c line for line:

    02007008  bl 0x2059788    func_02059788()
    0200700c  bl 0x2059bc0    func_02059bc0()
    02007018  bl 0x2059578    func_02059578(&func_0201a03c, &data_020a621c)
    02007028  bl 0x203c408    Heap::SetNodeID(*data_020a0e9c, 2)
    0200702c  bl 0x201a054    func_0201a054()
    0200703c  bl 0x203c408    Heap::SetNodeID(*data_020a0e9c, 3)
    02007040  bl 0x20197b8    func_020197b8()

Both seam calls are BL instructions in main's own compiled body and nothing
takes their address, so neither seam is a keep-alive: the ROM's main really
makes both calls. And the host calls it where the ROM calls it. Entry
(0x02004800) runs func_02019780 at 0x020048a4 and func_02072f94 at 0x020048a8,
then loads main's address from its own literal pool (0x020048d0 = 0x02007000)
and a pooled lr of 0xffff0000 and TAIL-JUMPS with `bx r1`. tests/walk_window.cpp
calls port_rom_main_run() immediately after port_boot_rom_pre_main() and
Heap::InitializeRootHeap(), which are func_02019780's own body and its fourth
call -- the same point.

Neither func_0201a054 nor func_020197b8 is enrolled on any slice list, so the
per-source rename claims nothing and invents no shadow row. The two words
main.c hands func_02059578 are the same two the transcription handed it:
hal/boot_os.cpp defines data_020a621c as char[348], so its bare name and
main.c's &data_020a621c are one address.

hal/rom_main_scout.cpp is renamed hal/rom_main.cpp: it is the port's boot spine
now, not a scout's measurement, and its header block is rewritten to say so.
The CMake block moves above the CONFIGURE-DEPENDS self-check so that report
stays the last thing the file does; it reads no lists, so the registered-input
and raw-read counts do not move.

THE ONE KNOWN DEVIATION IS UNCHANGED AND STATED. The ROM's order is head,
SetNodeID(2), func_0201a054, SetNodeID(3), loop, and the port still runs
func_0201a054's two transcribed spans further down tests/walk_window.cpp, so
the second SetNodeID happens before those spans instead of after them. Both
calls happen now, where before neither did.

AND THE PREVIOUS COMMIT IS REBUILT. BOOTSCOUT's smoke_player line landed
unbuilt because its session ended; this lane rebuilt the whole tree with it, so
all three targets link and the battery it was blocking has now run.

MEASURED on this tree, gate at its new default:

    linked into walk_window : 10145 -> 10147 (89.6%), +2
      main.c.obj                      defines _rom_main
      _ZN4Heap9SetNodeIDEj.cpp.obj    defines ?SetNodeID@Heap@@QAEXI@Z
    UNDOCUMENTED QUEUE / SHADOWS : 15 -> 14, no new shadow row
    CONFIGURE DEPS 246 inputs, 5 raw reads
    closestplayer_guard OK, inferred_stub_guard OK (no new seat),
    alternatename_guard OK (0 new defeats), gxband layout OK,
    tailjump forms OK
